+++ tsfile/src/test/java/org/apache/iotdb/tsfile/read/TsFileSequenceReaderTest.java
@@ -29,15 +29,12 @@ import java.util.Map.Entry;
 import java.util.Set;
 import java.util.stream.Collectors;
 import org.apache.iotdb.tsfile.common.conf.TSFileConfig;
-import org.apache.iotdb.tsfile.common.conf.TSFileDescriptor;
 import org.apache.iotdb.tsfile.file.MetaMarker;
 import org.apache.iotdb.tsfile.file.footer.ChunkGroupFooter;
 import org.apache.iotdb.tsfile.file.header.ChunkHeader;
 import org.apache.iotdb.tsfile.file.header.PageHeader;
 import org.apache.iotdb.tsfile.file.metadata.ChunkMetadata;
-import org.apache.iotdb.tsfile.file.metadata.TimeseriesMetadata;
 import org.apache.iotdb.tsfile.read.common.Chunk;
-import org.apache.iotdb.tsfile.read.common.Path;
 import org.apache.iotdb.tsfile.utils.FileGenerator;
 import org.apache.iotdb.tsfile.utils.Pair;
 import org.junit.After;
@@ -49,14 +46,10 @@ public class TsFileSequenceReaderTest {
 
   private static final String FILE_PATH = FileGenerator.outputDataFile;
   private ReadOnlyTsFile tsFile;
-  private final TSFileConfig conf = TSFileDescriptor.getInstance().getConfig();
-  private int maxDegreeOfIndexNode;
 
   @Before
   public void before() throws IOException {
     int rowCount = 100;
-    maxDegreeOfIndexNode = conf.getMaxDegreeOfIndexNode();
-    conf.setMaxDegreeOfIndexNode(3);
     FileGenerator.generateFile(rowCount, 10000);
     TsFileSequenceReader fileReader = new TsFileSequenceReader(FILE_PATH);
     tsFile = new ReadOnlyTsFile(fileReader);
@@ -66,7 +59,6 @@ public class TsFileSequenceReaderTest {
   public void after() throws IOException {
     tsFile.close();
     FileGenerator.after();
-    conf.setMaxDegreeOfIndexNode(maxDegreeOfIndexNode);
   }
 
   @Test
@@ -161,36 +153,5 @@ public class TsFileSequenceReaderTest {
     // test for non-exist device "d3"
     Assert.assertTrue(reader.readChunkMetadataInDevice("d3").isEmpty());
     reader.close();
-  }
-
-  @Test
-  public void testReadTimeseriesMetadata() throws IOException {
-    TsFileSequenceReader reader = new TsFileSequenceReader(FILE_PATH);
-    Path path = new Path("d1", "s1");
-    Set<String> set = new HashSet<>();
-    set.add("s1");
-    set.add("s2");
-    set.add("s3");
-    // the Max Degree Of Index Node is set to be 3, so the leaf node should only contains 3 sensors
-    // s4 should not be returned as result
-    set.add("s4");
-    List<TimeseriesMetadata> timeseriesMetadataList = reader.readTimeseriesMetadata(path, set);
-    Assert.assertEquals(3, timeseriesMetadataList.size());
-    for (int i = 1; i <= timeseriesMetadataList.size(); i++) {
-      Assert.assertEquals("s" + i, timeseriesMetadataList.get(i - 1).getMeasurementId());
-    }
-
-    path = new Path("d1", "s5");
-    set.clear();
-    set.add("s5");
-    set.add("s6");
-    // this is a fake one, this file doesn't contain this measurement
-    // so the result is not supposed to contain this measurement's timeseries metadata
-    set.add("s8");
-    timeseriesMetadataList = reader.readTimeseriesMetadata(path, set);
-    Assert.assertEquals(2, timeseriesMetadataList.size());
-    for (int i = 5; i < 7; i++) {
-      Assert.assertEquals("s" + i, timeseriesMetadataList.get(i - 5).getMeasurementId());
-    }
   }
 }
